requirements.txt
ib_insync
yfinance
pandas
python-dotenv